Move hook and rule to dedicated pkgs
