+++ internal/auth0/rule/resource_config_test.go
@@ -1,4 +1,4 @@
-package provider
+package rule_test
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/sweep"
 	"github.com/auth0/terraform-provider-auth0/internal/template"
@@ -15,11 +16,45 @@ func init() {
 	sweep.RuleConfigs()
 }
 
+// This is needed so that the test
+// sweepers get registered.
+func TestMain(m *testing.M) {
+	resource.TestMain(m)
+}
+
+const testAccRuleConfigCreate = `
+resource "auth0_rule_config" "foo" {
+  key = "acc_test_{{.testName}}"
+  value = "bar"
+}
+`
+
+const testAccRuleConfigUpdateValue = `
+resource "auth0_rule_config" "foo" {
+  key = "acc_test_{{.testName}}"
+  value = "foo"
+}
+`
+
+const testAccRuleConfigUpdateKey = `
+resource "auth0_rule_config" "foo" {
+  key = "acc_test_key_{{.testName}}"
+  value = "foo"
+}
+`
+
+const testAccRuleConfigEmptyValue = `
+resource "auth0_rule_config" "foo" {
+  key = "acc_test_key_{{.testName}}"
+  value = ""
+}
+`
+
 func TestAccRuleConfig(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: TestFactories(httpRecorder),
+		ProviderFactories: provider.TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccRuleConfigCreate, t.Name()),
@@ -56,31 +91,3 @@ func TestAccRuleConfig(t *testing.T) {
 		},
 	})
 }
-
-const testAccRuleConfigCreate = `
-resource "auth0_rule_config" "foo" {
-  key = "acc_test_{{.testName}}"
-  value = "bar"
-}
-`
-
-const testAccRuleConfigUpdateValue = `
-resource "auth0_rule_config" "foo" {
-  key = "acc_test_{{.testName}}"
-  value = "foo"
-}
-`
-
-const testAccRuleConfigUpdateKey = `
-resource "auth0_rule_config" "foo" {
-  key = "acc_test_key_{{.testName}}"
-  value = "foo"
-}
-`
-
-const testAccRuleConfigEmptyValue = `
-resource "auth0_rule_config" "foo" {
-  key = "acc_test_key_{{.testName}}"
-  value = ""
-}
-`
