+++ internal/auth0/rule/resource_test.go
@@ -1,12 +1,12 @@
-package provider
+package rule_test
 
 import (
 	"fmt"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/template"
 )
@@ -15,7 +15,7 @@ func TestAccRule(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: TestFactories(httpRecorder),
+		ProviderFactories: provider.TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(testAccRuleCreate, t.Name()),
@@ -71,25 +71,3 @@ resource "auth0_rule" "my_rule" {
   enabled = false
 }
 `
-
-func TestRuleNameRegexp(t *testing.T) {
-	vf := validation.StringMatch(ruleNameRegexp, "invalid name")
-
-	for name, valid := range map[string]bool{
-		"my-rule-1":                 true,
-		"1-my-rule":                 true,
-		"rule 2 name with spaces":   true,
-		" rule with a space prefix": false,
-		"rule with a space suffix ": false,
-		" ":                         false,
-		"   ":                       false,
-	} {
-		_, errs := vf(name, "name")
-		if errs != nil && valid {
-			t.Fatalf("Expected %q to be valid, but got validation errors %v", name, errs)
-		}
-		if errs == nil && !valid {
-			t.Fatalf("Expected %q to be invalid, but got no validation errors.", name)
-		}
-	}
-}
